# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HapiControl Internal Financial Operations API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List clients
      parameters:
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: executiveId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Clients list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get client by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClientBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}/risk-score:
    get:
      operationId: getClientRiskScore
      tags: [clients]
      summary: Get client risk score
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Risk score
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskScore"

  # CREDITS
...
  /commitments:
    get:
      operationId: listCommitments
      tags: [commitments]
      summary: List payment commitments
      parameters:
        - name: clientId
          in: query
          schema:
            type: ["integer", "null"]
        - name: executiveId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Commitments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Commitment"
    post:
      operationId: createCommitment
      tags: [commitments]
      summary: Create payment commitment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommitmentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Commitment"
  /commitments/{id}:
    patch:
      operationId: updateCommitment
      tags: [commitments]
      summary: Update commitment status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCommitmentBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Commitment"

  # NOTES
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Client:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        phone:
          type: string
        altPhone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        curp:
          type: ["string", "null"]
        status:
          type: string
          enum: [current, at_risk, overdue, defaulted, inactive]
        executiveId:
          type: ["integer", "null"]
        executiveName:
          type: ["string", "null"]
        riskLevel:
          type: ["string", "null"]
        registeredAt:
          type: string
        updatedAt:
          type: string
      required: [id, fullName, phone, status, registeredAt, updatedAt]
    ClientDetail:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        phone:
          type: string
        altPhone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        curp:
          type: ["string", "null"]
        status:
          type: string
          enum: [current, at_risk, overdue, defaulted, inactive]
        executiveId:
          type: ["integer", "null"]
        executiveName:
          type: ["string", "null"]
        riskLevel:
          type: ["string", "null"]
        guarantorName:
          type: ["string", "null"]
        guarantorPhone:
          type: ["string", "null"]
        internalNotes:
          type: ["string", "null"]
        registeredAt:
          type: string
        updatedAt:
          type: string
        credits:
          type: array
          items:
            $ref: "#/components/schemas/Credit"
        recentPayments:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
        recentNotes:
          type: array
          items:
            $ref: "#/components/schemas/Note"
        openCommitments:
          type: array
          items:
            $ref: "#/components/schemas/Commitment"
      required: [id, fullName, phone, status, registeredAt, updatedAt, credits, recentPayments, recentNotes, openCommitments]
    CreateClientBody:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        altPhone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        curp:
          type: ["string", "null"]
        executiveId:
          type: ["integer", "null"]
        guarantorName:
          type: ["string", "null"]
        guarantorPhone:
          type: ["string", "null"]
        internalNotes:
          type: ["string", "null"]
      required: [fullName, phone]
    UpdateClientBody:
      type: object
      properties:
        fullName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        altPhone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        curp:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        executiveId:
          type: ["integer", "null"]
        guarantorName:
          type: ["string", "null"]
        guarantorPhone:
          type: ["string", "null"]
        internalNotes:
          type: ["string", "null"]
    RiskScore:
      type: object
      properties:
        clientId:
          type: integer
        score:
          type: integer
        level:
          type: string
          enum: [low, medium, high]
        punctualityScore:
          type: integer
        delayCount:
          type: integer
        brokenPromises:
          type: integer
        currentArrears:
          type: number
        renewalCount:
          type: integer
        calculatedAt:
          type: string
      required: [clientId, score, level, punctualityScore, delayCount, brokenPromises, currentArrears, renewalCount, calculatedAt]
    Credit:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: ["string", "null"]
        executiveId:
          type: ["integer", "null"]
        executiveName:
          type: ["string", "null"]
        amount:
          type: number
        disbursementDate:
          type: string
        termWeeks:
          type: integer
        weeklyPayment:
          type: number
        openingFee:
          type: ["number", "null"]
        totalToRepay:
          type: number
        currentPaymentNumber:
          type: integer
        remainingBalance:
          type: number
        status:
          type: string
          enum: [active, completed, defaulted, restructured]
        renewalEligible:
          type: boolean
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, clientId, amount, disbursementDate, termWeeks, weeklyPayment, totalToRepay, currentPaymentNumber, remainingBalance, status, renewalEligible, createdAt, updatedAt]
...
    Payment:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: ["string", "null"]
        creditId:
          type: integer
        paymentNumber:
          type: integer
        paymentDate:
          type: string
        amountPaid:
          type: number
        amountExpected:
          type: number
        lateFee:
          type: ["number", "null"]
        updatedBalance:
          type: number
        paymentStatus:
          type: string
          enum: [on_time, partial, late, missed]
        executiveId:
          type: ["integer", "null"]
        executiveName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, clientId, creditId, paymentNumber, paymentDate, amountPaid, amountExpected, updatedBalance, paymentStatus, createdAt]
...
    Commitment:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: ["string", "null"]
        executiveId:
          type: ["integer", "null"]
        executiveName:
          type: ["string", "null"]
        promisedDate:
          type: string
        promisedAmount:
          type: number
        reason:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, fulfilled, broken]
        followUpNote:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, clientId, promisedDate, promisedAmount, status, createdAt, updatedAt]
    CreateCommitmentBody:
      type: object
      properties:
        clientId:
          type: integer
        executiveId:
          type: ["integer", "null"]
        promisedDate:
          type: string
        promisedAmount:
          type: number
        reason:
          type: ["string", "null"]
      required: [clientId, promisedDate, promisedAmount]
    UpdateCommitmentBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, fulfilled, broken]
        followUpNote:
          type: ["string", "null"]
      required: [status]
    Note:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        authorId:
          type: ["integer", "null"]
        authorName:
          type: ["string", "null"]
        noteType:
          type: string
          enum: [call, visit, comment, promise, issue, other]
        content:
          type: string
        createdAt:
          type: string
      required: [id, clientId, noteType, content, createdAt]
...